+++ src/main/java/io/github/syst3ms/skriptparser/expressions/ExprLength.java
@@ -5,6 +5,7 @@ import io.github.syst3ms.skriptparser.lang.TriggerContext;
 import io.github.syst3ms.skriptparser.lang.base.PropertyExpression;
 import org.jetbrains.annotations.Nullable;
 
+import java.math.BigInteger;
 import java.util.Optional;
 import java.util.function.Function;
 
@@ -31,7 +32,7 @@ public class ExprLength extends PropertyExpression<Number, String> {
 
     @Override
     public Optional<? extends Function<? super String[], ? extends Number[]>> getPropertyFunction() {
-        return Optional.of(strings -> new Number[]{strings[0].length()});
+        return Optional.of(strings -> new Number[]{BigInteger.valueOf(strings[0].length())});
     }
 
     @Override
